+++ .gitignore
@@ -1,18 +1,11 @@
 .gradle
-
 bin
-
 Reports
-
 Libs
-
 output
-
 !output/.gitkeep
 build
-
 .classpath
-
 .project
 /bin/
 /Libs/
